+++ Plugins/ReadyPlayerMe/Source/ReadyPlayerMe/Public/ReadyPlayerMeRenderLoader.h
@@ -4,7 +4,6 @@
 
 #include "CoreMinimal.h"
 #include "ReadyPlayerMeTypes.h"
-#include "Interfaces/IHttpRequest.h"
 #include "ReadyPlayerMeRenderLoader.generated.h"
 
 /**
@@ -22,16 +21,13 @@ public:
 	 * 
 	 * @param ModelUrl Model url.
 	 * @param SceneType The type of the scene where the avatar should be rendered.
-	 * @param Gender Gender is needed for rendering the correct pose.
 	 * @param OnCompleted Success callback. Called when the render is loaded and provides the avatar texture as an argument.
 	 * @param OnFailed Failure callback. If the render operation fails, the failure callback will be called.
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me", meta = (DisplayName = "Load", AutoCreateRefTerm = "OnCompleted, OnFailed"))
-	void Load(const FString& ModelUrl, const ERenderSceneType& SceneType, const EAvatarGender& Gender, const FDownloadImageCompleted& OnCompleted, const FDownloadImageFailed& OnFailed);
+	void Load(const FString& ModelUrl, const ERenderSceneType& SceneType, const FDownloadImageCompleted& OnCompleted, const FDownloadImageFailed& OnFailed);
 
 private:
-	void RequestRender(const FString& ModelUrl, const ERenderSceneType& SceneType, const EAvatarGender& Gender);
-
 	void DownloadRenderedImage(const FString& ImageUrl);
 
 	UFUNCTION()
@@ -39,8 +35,6 @@ private:
 
 	UFUNCTION()
 	void OnTexture2DDownloadFailed(UTexture2DDynamic* Texture);
-
-	void OnRenderCallback(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bSuccess);
 
 	UPROPERTY()
 	class UAsyncTaskDownloadImage* DownloadImageTask;